implemented rate limiter for http client and websocket client - keep in mind that the websocket client does not notify remaining weight so the only thing we can do is wait when the server blocks us - maybe in the future we can work with a client-side weight and sync with server-side when available

+++ src/main/java/be/davidopdebeeck/bitvavo/client/utils/ThreadUtils.java
@@ -4,7 +4,7 @@ import static java.lang.Thread.sleep;
 
 public class ThreadUtils {
 
-    public static void waitForMillis(int millis) {
+    public static void waitForMillis(long millis) {
         try {
             sleep(millis);
         } catch (InterruptedException e) {
